feat: Update delete button functionality in AdmPage.vue
The delete button in the AdmPage.vue component has been updated to call the deleteEmployee method when clicked. This method sends a DELETE request to the server to delete the employee with the specified ID. Additionally, the getEmployees method is called after a successful deletion to update the list of employees displayed on the page.